Adding Resource, Buffer, IndexBuffer, and VertexBuffer classes.
Adding Initialize function to Application class so that resource can be initialized after the Application singleton has been initialized. Some resource created by Application need access to the Application singleton. This means those resource couldn't be initialized in the constructor for the Application class (because the gs_Appliation global-static variable was not yet initialized.