Disable MSAA on HDR texture.
When scaling the HDR render target, the texture looked blocky after blitting to the SDR render target when using MSAA. Instead of MSAA, I'll implement FXAA on the final image (another time).
I also noticed an issue with the GenerateMips method when trying to generate the mipmaps for the HDR render target. This was an issue with the UAV flag being used on the UAV resource but not the alias resource (used to copy the original texture). The issue was fixed by using the UAV flag on both the UAV resource and placed resource.