Changed the way Update/Render work.
Note: There was an issue with the OpenFile Dialog that was caused by the update/render method that I was using.
The update/render functions are now separated which fixes the OpenFile dialog issue I was having.
But this causes other issues (like the window doesn't redraw while dragging the window)